Refactor du Session de Koala Rock et suppression de ce meme fichier
