+++ src/main/java/games/project/koala_rock/View/VueConnexion.java
@@ -5,7 +5,7 @@ import games.project.koala_rock.Metier.manager.PlayerManager;
 import games.project.koala_rock.Model.Fond;
 import games.project.koala_rock.RessourcesAccess;
 import games.project.stockage.Security;
-import games.project.koala_rock.Stockage.Session;
+import games.project.stockage.Session;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
