+++ src/main/java/games/project/koala_rock/View/VuePerdre.java
@@ -2,7 +2,7 @@ package games.project.koala_rock.View;
 
 import games.project.koala_rock.Model.Fond;
 import games.project.koala_rock.RessourcesAccess;
-import games.project.koala_rock.Stockage.Session;
+import games.project.stockage.Session;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
